requirements.txt
playwright
beautifulsoup4
lxml
requests
psycopg2-binary
python-dotenv
playwright-stealth
supabase
